requirements.txt
django-plotly-dash==2.3.1
fqdn==1.5.1
geopy==2.4.1
glob2==0.7
isoduration==20.11.0
jsonpointer==2.4
jupyter==1.0.0
matplotlib==3.9.4
mysqlclient==2.2.4
opencv-python==4.9.0.80
pandas==2.2.2
pip-chill==1.0.3
scikit-learn==1.5.1
tensorflow==2.18.0
tqdm==4.67.1
uri-template==1.3.0
webcolors==1.13
